# Run Ruby containers as root