--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -1,7 +1,7 @@
 version: '3.7'
 
 volumes:
-  home:
+  root-home:
   rabbitmq:
   postgres:
   mysql: