--- a/projects/bouncer/docker-compose.yml
+++ b/projects/bouncer/docker-compose.yml
@@ -16,7 +16,7 @@ services:
   bouncer-lite:
     <<: *bouncer
     depends_on:
-      - postgres
+      - postgres-9.6
     environment:
-      TEST_DATABASE_URL: "postgresql://postgres@postgres/transition_test"
+      TEST_DATABASE_URL: "postgresql://postgres@postgres-9.6/transition_test"
       DATABASE_CLEANER_ALLOW_REMOTE_DATABASE_URL: "true"