--- a/projects/cache-clearing-service/docker-compose.yml
+++ b/projects/cache-clearing-service/docker-compose.yml
@@ -7,7 +7,7 @@ x-cache-clearing-service: &cache-clearing-service
   image: cache-clearing-service
   volumes:
     - ${GOVUK_ROOT_DIR:-~/govuk}:/govuk:delegated
-    - home:/home/build
+    - root-home:/root
   working_dir: /govuk/cache-clearing-service
 
 services: