--- a/projects/calculators/docker-compose.yml
+++ b/projects/calculators/docker-compose.yml
@@ -9,14 +9,12 @@ x-calculators: &calculators
   tty: true
   volumes:
     - ${GOVUK_ROOT_DIR:-~/govuk}:/govuk:delegated
-    - home:/home/build
+    - root-home:/root
   working_dir: /govuk/calculators
 
 services:
   calculators-lite:
     <<: *calculators
-    # Tests need the privileged flag
-    privileged: true
 
   calculators-app: &calculators-app
     <<: *calculators