--- a/projects/calculators/docker-compose.yml
+++ b/projects/calculators/docker-compose.yml
@@ -1,5 +1,8 @@
 version: '3.7'
 
+volumes:
+  calculators-tmp:
+
 x-calculators: &calculators
   build:
     context: .
@@ -10,6 +13,7 @@ x-calculators: &calculators
   volumes:
     - ${GOVUK_ROOT_DIR:-~/govuk}:/govuk:delegated
     - root-home:/root
+    - calculators-tmp:/govuk/calculators/tmp
   working_dir: /govuk/calculators
 
 services: