--- a/projects/collections-publisher/docker-compose.yml
+++ b/projects/collections-publisher/docker-compose.yml
@@ -22,18 +22,18 @@ services:
   collections-publisher-lite:
     <<: *collections-publisher
     depends_on:
-      - mysql
+      - mysql-5.5
       - redis
     environment:
       # the app uses this URL to generate a test database URL
-      DATABASE_URL: "mysql2://root:root@mysql/collections_publisher_development"
+      DATABASE_URL: "mysql2://root:root@mysql-5.5/collections_publisher_development"
       REDIS_URL: redis://redis
       DATABASE_CLEANER_ALLOW_REMOTE_DATABASE_URL: "true"
 
   collections-publisher-app: &collections-publisher-app
     <<: *collections-publisher
     depends_on:
-      - mysql
+      - mysql-5.5
       - redis
       - content-store-app
       - link-checker-api-app
@@ -41,7 +41,7 @@ services:
       - publishing-api-app
       - collections-publisher-worker
     environment:
-      DATABASE_URL: "mysql2://root:root@mysql/collections_publisher_development"
+      DATABASE_URL: "mysql2://root:root@mysql-5.5/collections_publisher_development"
       REDIS_URL: redis://redis
       VIRTUAL_HOST: collections-publisher.dev.gov.uk
       HOST: 0.0.0.0
@@ -53,9 +53,9 @@ services:
     <<: *collections-publisher
     depends_on:
       - redis
-      - mysql
+      - mysql-5.5
       - publishing-api-app
     environment:
-      DATABASE_URL: "mysql2://root:root@mysql/collections_publisher_development"
+      DATABASE_URL: "mysql2://root:root@mysql-5.5/collections_publisher_development"
       REDIS_URL: redis://redis
     command: bundle exec sidekiq -C ./config/sidekiq.yml