--- a/projects/collections/docker-compose.yml
+++ b/projects/collections/docker-compose.yml
@@ -9,13 +9,12 @@ x-collections: &collections
   tty: true
   volumes:
     - ${GOVUK_ROOT_DIR:-~/govuk}:/govuk:delegated
-    - home:/home/build
+    - root-home:/root
   working_dir: /govuk/collections
 
 services:
   collections-lite:
     <<: *collections
-    privileged: true
 
   # TODO: doesn't work yet because collections also relies on search-api
   collections-app: &collections-app