--- a/projects/contacts-admin/docker-compose.yml
+++ b/projects/contacts-admin/docker-compose.yml
@@ -9,13 +9,12 @@ x-contacts-admin: &contacts-admin
   tty: true
   volumes:
     - ${GOVUK_ROOT_DIR:-~/govuk}:/govuk:delegated
-    - home:/home/build
+    - root-home:/root
   working_dir: /govuk/contacts-admin
 
 services:
   contacts-admin-lite:
     <<: *contacts-admin
-    privileged: true
     depends_on:
       - mysql
       - redis