--- a/projects/content-data-admin/docker-compose.yml
+++ b/projects/content-data-admin/docker-compose.yml
@@ -22,7 +22,7 @@ services:
   content-data-admin-lite:
     <<: *content-data-admin
     depends_on:
-      - postgres
+      - postgres-9.6
     environment:
-      DATABASE_URL: "postgresql://postgres@postgres/content-data-admin"
-      TEST_DATABASE_URL: "postgresql://postgres@postgres/content-data-admin-test"
+      DATABASE_URL: "postgresql://postgres@postgres-9.6/content-data-admin"
+      TEST_DATABASE_URL: "postgresql://postgres@postgres-9.6/content-data-admin-test"