--- a/projects/support-api/Dockerfile
+++ b/projects/support-api/Dockerfile
@@ -16,13 +16,6 @@ RUN curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | apt-key add -
 RUN echo "deb https://dl.yarnpkg.com/debian/ stable main" | tee /etc/apt/sources.list.d/yarn.list
 RUN apt-get update -qq && apt-get install -y yarn nodejs
 
-# Install phantomjs for some projects
-RUN export VERSION=2.1.1 && \
-    wget https://bitbucket.org/ariya/phantomjs/downloads/phantomjs-$VERSION-linux-x86_64.tar.bz2 2>&1 && \
-    tar -vxf phantomjs-$VERSION-linux-x86_64.tar.bz2 && \
-    cp -v phantomjs-$VERSION-linux-x86_64/bin/phantomjs /usr/local/bin/phantomjs && \
-    rm -vr phantomjs-$VERSION-linux-x86_64
-
 # Install rbenv to manage ruby versions
 RUN git clone https://github.com/sstephenson/rbenv.git /rbenv
 RUN git clone https://github.com/sstephenson/ruby-build.git /rbenv/plugins/ruby-build