Trello: https://trello.com/c/PlDX9ps8/180-run-govuk-docker-containers-as-root-to-allow-application-specific-mounts

This is a draft of an approach to run Ruby containers as root rather than the build user, this makes it easier to define volumes for individual projects ([this is rather difficult as a specified user](https://github.com/docker/compose/issues/3270)) which offers opportunities to improve SASSC performance substantially. I've opened this now to seek feedback on the approach before I go through and apply it to all the apps.

Applying this change will be a breaking change, which I need to have a think about. I'm wondering about planning to merge this into a 2.0 branch that can track master while we round up a few backwards incompatible changes that are needed. 

I did some benchmarks to show the difference in SASS performance, these are below and depend on projects using https://github.com/alphagov/govuk_test/pull/31. The tl;dr of these is that for content_publisher the application.css file compilation is sped up from 16s to 4s on a cold cache and 5s to 0.4s on a warm one, for govuk_publishing_components the application.css file compilation is sped up from 93s to 5s on a cold cache and 22s to 0.3s on a warm one.

<details>
  <summary>Benchmarks</summary>
  
## Content Publisher

### Rails Console

#### Before

Pre-steps:

```
cd ~/govuk/content-publisher
govuk-docker-run bundle exec rails tmp:clear
govuk-docker-run bundle exec rails console
```

```
irb(main):001:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  3.175072   1.573766   4.748838 ( 16.350221)
=> nil
irb(main):002:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.416941   0.503787   0.920728 (  5.863118)
=> nil
irb(main):003:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.532579   0.377048   0.909627 (  4.806736)
=> nil
irb(main):004:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.389468   0.400189   0.789657 (  4.725987)
=> nil
```

#### After

Pre-steps:

```
cd ~/govuk/content-publisher
govuk-docker-run bundle exec rails tmp:clear
govuk-docker-run bundle exec rails console
```

```
irb(main):001:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  2.513432   0.598473   3.111905 (  4.023456)
=> nil
irb(main):002:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.214788   0.053805   0.268593 (  0.307541)
=> nil
irb(main):003:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.299529   0.047013   0.346542 (  0.369209)
=> nil
irb(main):004:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.180049   0.050321   0.230370 (  0.253839)
=> nil
```

### Curling a component

#### Before

Pre-steps:

```
cd ~/govuk/content-publisher
govuk-docker-run bundle exec rails tmp:clear
govuk-docker-up
```

```
➜  ~  time curl -s -o /dev/null http://content-publisher.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 0% cpu 48.920 total
➜  ~  time curl -s -o /dev/null http://content-publisher.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 0% cpu 7.645 total
➜  ~  time curl -s -o /dev/null http://content-publisher.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 0% cpu 7.801 total
➜  ~  time curl -s -o /dev/null http://content-publisher.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 0% cpu 7.465 total
```

#### After

Pre-steps:

```
cd ~/govuk/content-publisher
govuk-docker-run bundle exec rails tmp:clear
govuk-docker-up
```

```
➜  ~ time curl -s -o /dev/null http://content-publisher.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 0% cpu 10.767 total
➜  ~ time curl -s -o /dev/null http://content-publisher.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 1% cpu 1.181 total
➜  ~ time curl -s -o /dev/null http://content-publisher.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 2% cpu 0.772 total
➜  ~ time curl -s -o /dev/null http://content-publisher.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 1% cpu 0.756 total
```

## `govuk_publishing_components` 

### bundle console

#### Before

```
$ cd ~/govuk/govuk_publishing_components
$ rm -rf spec/dummy/tmp
$ govuk-docker-run bundle console
```

```
irb(main):001:0> require_relative "spec/dummy/config/environment"
=> true
irb(main):002:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  4.103497   6.380649  10.484146 ( 93.095291)
=> nil
irb(main):003:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.662254   1.294566   1.956820 ( 20.607901)
=> nil
irb(main):004:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.662850   1.058770   1.721620 ( 21.924899)
=> nil
irb(main):005:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.707186   1.225398   1.932584 ( 21.971194)
=> nil
```

#### After

Pre-steps:

```
$ cd ~/govuk/govuk_publishing_components
$ docker volume rm govuk-docker_govuk_publishing_components_tmp
$ govuk-docker-run bundle console
```

```
irb(main):001:0> require_relative "spec/dummy/config/environment"
=> true
irb(main):002:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  2.264592   0.761231   3.025823 (  5.016340)
=> nil
irb(main):003:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.204888   0.062796   0.267684 (  0.376280)
=> nil
irb(main):004:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.187334   0.071032   0.258366 (  0.303015)
=> nil
irb(main):005:0> puts Benchmark.measure { Rails.application.assets["application.css"] }
  0.206238   0.056903   0.263141 (  0.434597)
=> nil
```

### Curling a component

#### Before

Pre-steps:

```
$ cd ~/govuk/govuk_publishing_components
$ rm -rf spec/dummy/tmp
$ govuk-docker-up
```

```
➜  ~ time curl -s -o /dev/null http://govuk-publishing-components.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 0% cpu 2:45.72 total
➜  ~ time curl -s -o /dev/null http://govuk-publishing-components.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 0% cpu 22.884 total
➜  ~ time curl -s -o /dev/null http://govuk-publishing-components.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 0% cpu 26.329 total
➜  ~ time curl -s -o /dev/null http://govuk-publishing-components.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 0% cpu 28.211 total
```

#### After

Pre-steps:

```
$ cd ~/govuk/govuk_publishing_components
$ docker volume rm govuk-docker_govuk_publishing_components_tmp
$ govuk-docker-up
```

```
➜  ~ time curl -s -o /dev/null http://govuk-publishing-components.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 0% cpu 12.945 total
➜  ~ time curl -s -o /dev/null http://govuk-publishing-components.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 1% cpu 1.078 total
➜  ~ time curl -s -o /dev/null http://govuk-publishing-components.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 1% cpu 0.729 total
➜  ~ time curl -s -o /dev/null http://govuk-publishing-components.dev.gov.uk/component-guide/button
curl -s -o /dev/null   0.01s user 0.01s system 1% cpu 0.717 total
```

</details>