@benthorner 2020-10-16
I'm trying this out as a beta tester. I went for option one and then built Email Alert * and ran tests for Static.

Only one issue: stopping all containers command had a parse error; used `docker ps -a -q | xargs docker rm -f` instead.