@kevindew 2020-10-16
> docker ps -a -q | xargs docker rm -f

Great thanks for that. That seems a better command anyway as it can handle no processes running which I means I can drop the conditional aspect.