@kevindew 2020-11-13
Thanks for the review Ben, lots of good points. An ADR does sound a good idea. 

node_modules should rather rapidly become the norm as I apply stylelint across the apps. I did have a hacky idea on how to test it, could use the presence of yarn in an applications makefile to suggest it should be there, does make me a little nauseous at the hackiness though but would probably work.