@kevindew 2020-11-16
@benthorner I've put a couple of tests in if you'd like to have a look. I'd not really looked at the other tests before and notice they're all rather circumstantial so the rather delicate node one I've put in feels consistent with existing approaches - I can see it's rather a challenge to work out where things are assertable.